backend/requirements.txt
djangorestframework
djangorestframework-simplejwt
gunicorn
psycopg2-binary
Django>=4.2,<5.0
django-cors-headers
